send mail when turn rejected && proper autentication in rejection

+++ serviYa/src/routes/turnos/+page.server.ts
@@ -1,8 +1,9 @@
 import { prisma } from '$lib/server/lucia/prisma';
 import { fail, redirect } from '@sveltejs/kit';
-import type { PageServerLoad, Actions } from './$types';
-import { notifyTurnAceptation } from '$lib/server/email/emailService';
+import type { PageServerLoad, Actions, RequestEvent, RequestEvent } from './$types';
+import { notifyTurnAceptation, notifyTurnRejection } from '$lib/server/email/emailService';
 import { EventEmitterAsyncResource } from 'nodemailer/lib/xoauth2';
+import type { RouteParams } from '../$types';
 
 export const load: PageServerLoad = async ({ locals }) => {
 	const { user } = await locals.auth.validateUser();
@@ -31,7 +32,9 @@ export const actions: Actions = {
 
 		const data = await request.formData();
 		const appointment_id = data.get('appointment_id')?.toString();
-
+		if (!appointment_id) {
+			return fail(400, { msj: 'No se envio el id del turno' });
+		}
 		const appointment = await prisma.appointment.findUnique({ where: { id: appointment_id } })
 		if (!appointment) {
 			return fail(404, { msj: 'Appointment not found' })
@@ -50,22 +53,37 @@ export const actions: Actions = {
 		prisma.authUser.findUnique({ where: { id: client_id } })
 			.then(client => {
 				if (!client) return
-				console.log('envio de mail')
 				notifyTurnAceptation({ to: client.email, date, profesionalName: user.name })
 			})
 		return { status: 200 };
 	},
 
-	reject: async ({ request }) => {
+	reject: async ({ request, locals }) => {
+		const { user } = await locals.auth.validateUser()
+
 		const data = await request.formData();
 		const appointment_id = data.get('appointment_id')?.toString();
-		//todo falta validar cosas (segun bruno)
 		if (!appointment_id) {
-			return Error('No se envio el id del turno');
+			return fail(400, { msj: 'No se envio el id del turno' });
 		}
+		const appointment = await prisma.appointment.findUnique({ where: { id: appointment_id } })
+		if (!appointment) {
+			return fail(404, { msj: 'Appointment not found' })
+		}
+		if (appointment.profesional_id != user?.userId) {
+			return fail(401, { msj: 'You lack the required credentilas' })
+		}
+
 		await prisma.appointment.delete({
 			where: { id: appointment_id }
 		});
+
+		const { client_id, date } = appointment
+		prisma.authUser.findUnique({ where: { id: client_id } })
+			.then(client => {
+				if (!client) return
+				notifyTurnRejection({ to: client.email, date, profesionalName: user.name })
+			})
 
 		return { status: 200, success: true };
 	}
